web: extract FightResultIcon
